AndroidManifest*.xml: Add yubico's url to VIEW/BROWSABLE intent-filter
This patch makes it so that the user can choose keepass2android as
a handler for my.yubico.com in
Settings->Apps->App links->keepass2android->Open supported links->
Open in this app.
This is needed to be set manually by the user in order for their
Yubikey NEOs to work.