Fix PCloud not overwriting uploaded files.
When uploading an existing file into PCloud, it appends a version number
by default (e.g. "MyDb (2).kdbx", "MyDb (3).kdbx" and so on). This
behavior is not desirable in K2PA case. The workaroundfor this is to upload
the file with a temporary name and then rename it to its final name.
This may also avoid corruption with failed uploads depending on how
PCloud handles uploads.

See https://github.com/PhilippC/keepass2android/issues/512#issuecomment-435475365