IMPROVED UNDERSCREEN FINGERPRINT, ISSUE #831
Improved underscreen fingerprint as request in issue #831 by giving the users
the ability to disable the sensor in the PasswordActivity. Now pressing the
fingerprint icon opens up a dialog that has the option to enable/disable the
sensor.
I added two new strings to the resources (enable sensor - disable sensor) and
already translated them to italian.