set AndroidClassParser build property to class-parse
Build reported these warnings:

warning XA4231: The Android class parser value 'jar2xml' is deprecated and will be removed in a future version of Xamarin.Android. Update the project properties to use 'class-parse'

In the code <AndroidClassParser> is not specified, so by default it uses
'jar2xml'. So set its value to 'class-parse'.

See:
https://github.com/xamarin/xamarin-android/blob/main/Documentation/guides/messages/xa4231.md
https://github.com/xamarin/xamarin-android/issues/1319

Moving to class-parse instead of jar2xml also fixes issues when building with Java 11 & Xamarin < 12.2
Because of:
- https://github.com/xamarin/xamarin-android/commit/94e21c7aaa8d5eb51e0e2234cd24d607515b0064
If the change wasn't done, we would have to stick with JDK-8