Rework SFTP path URI to put query string after path
Switches the order in which the URI is built to better align with a
typical URI.

Before this change, and example path URI looks like:
  sftp://host:port?extra=option1&another=option2/path/to/file
After this change, the same conceptual path looks like:
  sftp://host:port/path/to/file?extra=option1&another=option2

NOTE: This commit does NOT represent a working build (yet). There are
interoperability issues when passing the URIs to android-filechooser-AS,
since it does not understand the extra query string.