Integrate Sftp query param URIs with android-filechooser
Allow the full Uri to be used across the board (usually via FileEntry.path).
To make this work a few places where Uri manipulation occurs needed to be
changed in order to accept and handle the possibility of a query string.
This allows for full Uri consistency between SftpStorage and filechooser code.

Also expose query params in the display name, so that they show up in
Recents list (differentiate between configurations).